logging: Refactor in preparation for logging v2
Refactor and cleanup in preparation for v2.

+++ samples/subsys/logging/logger/src/main.c
@@ -69,15 +69,16 @@ static uint32_t timestamp_freq(void)
  *
  * @return Source ID.
  */
-static int log_source_id_get(const char *name)
+static int16_t log_source_id_get(const char *name)
 {
 
-	for (int i = 0; i < log_src_cnt_get(CONFIG_LOG_DOMAIN_ID); i++) {
+	for (int16_t i = 0; i < log_src_cnt_get(CONFIG_LOG_DOMAIN_ID); i++) {
 		if (strcmp(log_source_name_get(CONFIG_LOG_DOMAIN_ID, i), name)
 		    == 0) {
 			return i;
 		}
 	}
+
 	return -1;
 }
